requirements.txt
pip
Flask
cassandra-driver
requests
